Move common code to moduleDropdown.
